// backend/src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly PrismaMigrations: {
      readonly id: Varchar<36>;
      readonly checksum: Varchar<64>;
      readonly finishedAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly migrationName: Varchar<255>;
      readonly logs: CodecTypes['pg/text@1']['output'] | null;
      readonly rolledBackAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly startedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly appliedStepsCount: CodecTypes['pg/int4@1']['output'];
    };
    readonly Project: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly teamId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly nextTaskKey: CodecTypes['pg/int4@1']['output'];
      readonly authorId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Task: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'TODO' | 'IN_PROGRESS' | 'DONE';
      readonly priority: 'LOW' | 'MEDIUM' | 'HIGH';
      readonly projectId: CodecTypes['pg/int4@1']['output'];
      readonly assigneeId: CodecTypes['pg/int4@1']['output'] | null;
      readonly authorId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly keyNumber: CodecTypes['pg/int4@1']['output'];
    };
    readonly Team: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly ownerId: CodecTypes['pg/int4@1']['output'];
    };
    readonly TeamMember: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly teamId: CodecTypes['pg/int4@1']['output'];
      readonly role: 'ADMIN' | 'MEMBER';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
    };
  };
};